feat #5 CI Supports compiling the project
The CI Server now supports compiling the maven project using the command mvn clean package -DskipTests
Two test are added to try the functionality of CompileProject.java
Added test for the function CloneRepository.cloneRepository() as well.
Added in CIServer.java to handle JSON objects when a POST request is sent to the server.