feat #11 clone specific branch
Changes the Helpers class to add a function getBranch to extract the branch name and changes cloneRepository and calls to the function so that it takes the name and only clones the specified branch. Due to the changes, the tests for cloning the repostitory and compiling were updated. Therefore, no further tests were added.