fix #17 Added test for Testing
Renamed the function testBranch() to runTests().
A test for this function was created that clones, compiles and run the test
located in the path /test/java/group22.
This test however is located in the path /test/java/servertests which is not
executed when running "mvn test" in terminal. This configuration was added in
the pom.xml file. All tests are compiled but only tests located in /test/java/group22
are being executed.