fix #18 Runs group22 tests as default
When using the command "mvn test" it runs the tests located in the folder
test/java/group22 . A profile is created in the pom.xml file so to test the
actual server you now type "mvn test -Pservertest" and this will instead run the
tests located in the folder test/java/servertests.
All tests in /servertests are now contained in the package servertests;
this was not the case prior.